fix: resolve IPC pipe deadlock + add multi-process observability logging (#19)
DiffTracker's fire-and-forget `tokio::spawn + emit` held the transport
writer mutex while flushing, blocking the subsequent AwaitingInput emit
when the pipe was under pressure. Replaced with `try_emit()` which
spawns independently.

Added defensive `interrupt.take()` after every `wait_for_input()` return
to prevent stale IPC interrupt signals from pre-cancelling new turns.

Comprehensive debug/info logging across the entire IPC pipeline:
- connection.rs: send_request, send_notification, respond (method+id)
- bridge.rs: message forwarding, permission decisions, request routing
- ipc_frontend.rs: recv_input, request_permission, ask_user
- server.rs: agent start params (cwd, model, mode)
- cancel.rs: pre-cancel detection
- turn_exec.rs: turn cancellation
- loop_detector.rs: loop warnings/aborts
- event_handler.rs: AwaitingInput/Interrupted state transitions
- inbox.rs: forwarding decisions
- controller.rs: interrupt signals
- key_dispatch.rs: message submit routing
- stdio.rs: transport disconnect warnings

Split oversized files: bridge.rs → bridge_handlers.rs,
ipc_frontend.rs → ipc_emitter.rs.
