+++ crates/loopal-agent-client/src/bridge.rs
@@ -4,17 +4,15 @@
 //! creating a second reader loop on the same Transport.
 
 use std::sync::Arc;
-use std::time::Duration;
 
 use tokio::sync::mpsc;
-use tracing::{info, warn};
+use tracing::{debug, info, warn};
 
 use loopal_ipc::connection::{Connection, Incoming};
 use loopal_ipc::protocol::methods;
 use loopal_protocol::{AgentEvent, ControlCommand, Envelope, UserQuestionResponse};
 
-/// Timeout for permission/question responses from TUI (prevents infinite hang).
-const RESPONSE_TIMEOUT: Duration = Duration::from_secs(300);
+use crate::bridge_handlers::{handle_permission, handle_question};
 
 /// Handles for the TUI side of the IPC bridge.
 pub struct BridgeHandles {
@@ -71,6 +69,7 @@ pub fn start_bridge(
     let conn_msg = connection.clone();
     tokio::spawn(async move {
         while let Some(envelope) = mailbox_rx.recv().await {
+            debug!(target_agent = %envelope.target, "bridge: forwarding message");
             if let Ok(params) = serde_json::to_value(&envelope) {
                 if let Err(e) = conn_msg
                     .send_request(methods::AGENT_MESSAGE.name, params)
@@ -121,6 +120,7 @@ async fn bridge_incoming(
                 }
             }
             Incoming::Request { id, method, params } => {
+                debug!(id, %method, "bridge: incoming request");
                 if method == methods::AGENT_PERMISSION.name {
                     handle_permission(&connection, &event_tx, permission_rx, id, params).await;
                 } else if method == methods::AGENT_QUESTION.name {
@@ -137,87 +137,4 @@ async fn bridge_incoming(
             }
         }
     }
-}
-
-async fn handle_permission(
-    connection: &Connection,
-    event_tx: &mpsc::Sender<AgentEvent>,
-    permission_rx: &mut mpsc::Receiver<bool>,
-    request_id: i64,
-    params: serde_json::Value,
-) {
-    let tool_name = params["tool_name"]
-        .as_str()
-        .unwrap_or("unknown")
-        .to_string();
-    let tool_input = params.get("tool_input").cloned().unwrap_or_default();
-    let tool_id = params["tool_call_id"].as_str().unwrap_or("").to_string();
-    let event = AgentEvent {
-        agent_name: None,
-        payload: loopal_protocol::AgentEventPayload::ToolPermissionRequest {
-            id: tool_id,
-            name: tool_name,
-            input: tool_input,
-        },
-    };
-    let _ = event_tx.send(event).await;
-    // Wait with timeout — prevents infinite hang if TUI disappears
-    let allow = match tokio::time::timeout(RESPONSE_TIMEOUT, permission_rx.recv()).await {
-        Ok(Some(v)) => v,
-        _ => {
-            warn!("permission response timeout/closed, denying");
-            false
-        }
-    };
-    let _ = connection
-        .respond(request_id, serde_json::json!({"allow": allow}))
-        .await;
-}
-
-async fn handle_question(
-    connection: &Connection,
-    event_tx: &mpsc::Sender<AgentEvent>,
-    question_rx: &mut mpsc::Receiver<UserQuestionResponse>,
-    request_id: i64,
-    params: serde_json::Value,
-) {
-    let parsed = serde_json::from_value(params.get("questions").cloned().unwrap_or_default());
-    if let Ok(questions) = parsed {
-        let event = AgentEvent {
-            agent_name: None,
-            payload: loopal_protocol::AgentEventPayload::UserQuestionRequest {
-                id: "ipc".into(),
-                questions,
-            },
-        };
-        let _ = event_tx.send(event).await;
-    } else {
-        // Parse failed — respond immediately instead of waiting 300s
-        warn!("IPC bridge: failed to parse questions, auto-responding");
-        let fallback = UserQuestionResponse {
-            answers: vec!["(parse error)".into()],
-        };
-        let _ = connection
-            .respond(
-                request_id,
-                serde_json::to_value(&fallback).unwrap_or_default(),
-            )
-            .await;
-        return;
-    }
-    let response = match tokio::time::timeout(RESPONSE_TIMEOUT, question_rx.recv()).await {
-        Ok(Some(v)) => v,
-        _ => {
-            warn!("question response timeout/closed");
-            UserQuestionResponse {
-                answers: vec!["(timeout)".into()],
-            }
-        }
-    };
-    let _ = connection
-        .respond(
-            request_id,
-            serde_json::to_value(&response).unwrap_or_default(),
-        )
-        .await;
 }
