+++ crates/loopal-session/Cargo.toml
@@ -15,3 +15,4 @@ indexmap = { workspace = true }
 serde_json = { workspace = true }
 tokio = { workspace = true }
 chrono = { workspace = true }
+tracing = { workspace = true }
